// src/stores/models/enums/enums.ts
export enum AnalyticType {
  FIGHT = "FIGHT",
  LIST = "LIST",
  RS = "RS",
  UP = "UP",
}

export enum CornerType {
  FANTASY = "FANTASY",
  HISTORICAL = "HISTORICAL",
  LIVE = "LIVE",
}

export enum DistanceMetasType {
  JUDGE = "JUDGE",
  PROPS = "PROPS",
  PREDICTIONS = "PREDICTIONS",
}

export enum DistanceType {
  FIGHT = "FIGHT",
  ME = "ME",
  SHOW = "SHOW",
  SEASON = "SEASON",
  TEMP = "TEMP",
}

export enum JudgeType {
  COACH = "COACH",
  FSL = "FSL",
  OFFICIAL = "OFFICIAL",
}

export enum InviteType {
  CORNER_ID = "CORNER_ID",
  DISTANCE_ID = "DISTANCE_ID",
  EMAIL = "EMAIL",
}

export enum Networks {
  BLKPRIME = "BLKPRIME",
  DAZN = "DAZN",
  ESPN = "ESPN",
  ESPNPLUS = "ESPN+",
  FSL = "FSL",
  FIGHTTV = "FIGHT TV",
  FITETV = "FITE TV",
  HBO = "HBO",
  HBOPPV = "HBO PPV",
  SHOWTIME = "ShowTime",
  SHOWTIMEPPV = "ShowTime PPV",
  NONE = "NONE",
}

export enum OfficialResultType {
  OFFICIAL_JUDGES = "OFFICIAL_JUDGES",
  FSL_JUDGES = "FSL_JUDGES",
  COACHES = "COACHES",
}

export enum SeasonType {
  ANNUAL = "ANNUAL",
  FANTASY = "FANTASY",
  HISTORICAL = "HISTORICAL",
  MONTH = "MONTH",
  QUARTER = "QUARTER",
  TRAINING = "TRAINING",
}

export enum ShowType {
  AMATEUR = "AMATEUR",
  PROFESSIONAL = "PROFESSIONAL",
}

export enum Status {
  ACTIVE = "ACTIVE",
  ARCHIVED = "ARCHIVED",
  CANCELED = "CANCELED",
  COMPLETE = "COMPLETE",
  PENDING = "PENDING",
  TESTING = "TESTING",
}

export enum UserPickType {
  FIGHT = "FIGHT",
  LIST = "LIST",
}

export enum WeightClass {
  HEAVYWEIGHT = "Heavyweight",
  CRUISERWEIGHT = "Cruiserweight",
  LIGHTHEAVYWEIGHT = "Light Heavyweight",
  SUPERMIDDLEWEIGHT = "Super Middleweight",
  MIDDLEWEIGHT = "Middleweight",
  SUPERWELTERWEIGHT = "Super Welterweight",
  WELTERWEIGHT = "Welterwieght",
  JRWELTERWEIGHT = "Jr Welterweight",
  SUPERLIGHTWEIGHT = "Super Lightweight",
  LIGHTWEIGHT = "Lightweight",
  SUPERFEATHERWEIGHT = "Super Featherweight",
  FEATHERWEIGHT = "Featherweight",
  SUPERBANTAMWEIGHT = "Super Bantamweight",
  BANTAMWEIGHT = "Bantamweight",
  SUPERFLYWEIGHT = "Super Flyweight",
  FLYWEIGHT = "Flyweight",
  JRFLYWEIGHT = "Jr Flyweight",
}